refactored except for the error functions
