+++ config/panda.yaml
@@ -1,0 +1,40 @@
+dh_parameter:
+  joint1: [0, 0.333, 0, 0]
+  joint2: [0, 0, 0, -1.57079633]
+  joint3: [0, 0.316, 0, 1.57079633]
+  joint4: [0, 0, 0.0825, 1.57079633]
+  joint5: [0, 0.384, -0.0825, -1.57079633]
+  joint6: [0, 0, 0, 1.57079633]
+  joint7: [0, 0, 0.088, 1.57079633]
+
+su_dh_parameter:
+  su1: [1.57079633, -0.15, -1.57079633, 0.05, 0, 1.57079633]
+  su2: [1.57079633, 0.06, -1.57079633, 0.06, 0, 1.57079633]
+  su3: [0, -0.08, 0, 0.05, 0, 1.57079633]
+  su4: [-1.57079633, 0.08, 1.57079633, 0.06, 0, 1.57079633]
+  su5: [3.14, -0.1, 3.14, 0.1, 0, 1.57079633]
+  su6: [-1.57079633, 0.03, 1.57079633, 0.05, 0, 1.57079633]
+  su7: [1.57079633, 0, -1.57079633, 0.05, 0, 1.57079633]
+
+su_poses:
+  su1:
+    position: [0.050, -0.000, 0.183]
+    rotation: [0.000, 0.707, -0.000, 0.707]
+  su2:
+    position: [0.060, 0.060, 0.333]
+    rotation: [-0.500, 0.500, -0.500, 0.500]
+  su3:
+    position: [0.000, -0.050, 0.569]
+    rotation: [0.707, 0.000, -0.000, 0.707]
+  su4:
+    position: [0.023, -0.080, 0.653]
+    rotation: [0.482, -0.482, -0.517, 0.517]
+  su5:
+    position: [0.020, 0.100, 0.938]
+    rotation: [-0.706, 0.025, 0.025, 0.707]
+  su6:
+    position: [-0.023, -0.030, 1.041]
+    rotation: [0.482, -0.482, -0.517, 0.517]
+  su7:
+    position: [0.165, 0.000, 1.028]
+    rotation: [0.732, 0.000, 0.682, -0.000]
