+++ config/sawyer.yaml
@@ -1,0 +1,41 @@
+dh_parameter:
+  joint1: [0, 0.317, 0, 0]
+  joint2: [1.57079633, 0.1925, 0.081, -1.57079633]
+  joint3: [0, 0.4, 0, 1.57079633]
+  joint4: [0, -0.1685, 0, -1.57079633]
+  joint5: [0, 0.4, 0, 1.57079633]
+  joint6: [0, 0.1363, 0, -1.57079633]
+  joint7: [1.57079633, 0.13375, 0, 1.57079633]
+
+su_dh_parameter:
+  su1: [1.57079633, -0.157, -1.57079633, 0.07, 0, 1.57079633]
+  su2: [-1.57079633, -0.0925, 1.57079633, 0.07, 0, 1.57079633]
+  su3: [-1.57079633, -0.16, 1.57079633, 0.05, 0, 1.57079633]
+  su4: [-1.57079633, 0.0165, 1.57079633, 0.05, 0, 1.57079633]
+  su5: [-1.57079633, -0.17, 1.57079633, 0.04, 0, 1.57079633]
+  su6: [-1.57079633, 0.0053, 1.57079633, 0.04, 0, 1.57079633]
+  su7: [0.0, 0.12375, 0.0, 0.03, 0, -1.57079633]
+
+su_poses:
+  su1:
+    position: [0.070, -0.000, 0.16]
+    rotation: [-0.000, 0.707, 0.000, 0.707]
+  su2:
+    position: [0.086, 0.100, 0.387]
+    rotation: [0.024, 0.025, 0.707, 0.707]
+  su3:
+    position: [0.324, 0.191, 0.350]
+    rotation: [0.012, 0.035, -0.000, 0.999]
+  su4:
+    position: [0.485, 0.049, 0.335]
+    rotation: [0.045, 0.029, 0.706, 0.706]
+  su5:
+    position: [0.709, 0.023, 0.312]
+    rotation: [0.008, 0.052, -0.000, 0.999]
+  su6:
+    position: [0.883, 0.154, 0.287]
+    rotation: [0.045, 0.034, 0.706, 0.706]
+  su7:
+    position: [1.087, 0.131, 0.228]
+    rotation: [0.489, -0.428, 0.512, 0.562]
+
