+++ requirements.txt
@@ -11,4 +11,5 @@ nlopt
 smbus2
 VL53L1X
 pyquaternion
+pyyaml
 flake8
